Refactoring Repository layer
