+++ project-domain/src/main/java/by/farshyniou/jdbc/repository/impl/JdbcRepositoryImpl.java
@@ -35,194 +35,7 @@ public class JdbcRepositoryImpl implements JdbcRepository {
     }
 
     @Override
-    public void createCatTable() {
-        try (Connection connection = JdbcConnection.getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.execute(CREATE_TABLE_CAT);
-        } catch (SQLException e) {
-            LOGGER.error("Error during creating Cat table, {}", e.getMessage());
-        }
-    }
-
-    @Override
-    public void createBreedTable() {
-        try (Connection connection = JdbcConnection.getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.execute(CREATE_TABLE_BREED);
-        } catch (SQLException e) {
-            LOGGER.error("Error during creating Breed table, {}", e.getMessage());
-        }
-    }
-
-    @Override
-    public void createTables() {
-        try (Connection connection = JdbcConnection.getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.execute(CREATE_TABLE_BREED);
-            statement.execute(CREATE_TABLE_CAT);
-        } catch (SQLException e) {
-            LOGGER.error("Error during creating tables, {}", e.getMessage());
-        }
-    }
-
-    @Override
-    public void insertIntoTablesExample() {
-        try (Connection connection = JdbcConnection.getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.executeUpdate(Queries.INSERT_INTO_BREED_EXAMPLE);
-            statement.executeUpdate(Queries.INSERT_INTO_CAT_EXAMPLE);
-        } catch (SQLException e) {
-            LOGGER.error("Error during inserting into tables example values, {}", e.getMessage());
-        }
-    }
-
-    @Override
-    public Long deleteFromCatTable(Long catId) {
-        try (Connection connection = JdbcConnection.getConnection();
-             PreparedStatement pStatement = connection.prepareStatement(Queries.DELETE_FROM_CAT, Statement.RETURN_GENERATED_KEYS)) {
-            LOGGER.info("Deleting Cat with id {} ...", catId);
-            pStatement.setLong(1, catId);
-            pStatement.executeUpdate();
-            ResultSet keys = pStatement.getGeneratedKeys();
-            if (keys.next()) {
-                return keys.getLong(1);
-            } else {
-                LOGGER.info("Nothing to delete from cat for id {}", catId);
-            }
-
-        } catch (SQLException exception) {
-            LOGGER.error("Exception during deleting from Cat table, {}", exception.getMessage());
-            throw new RepositoryException(exception.getMessage());
-        }
-        return null;
-    }
-
-    @Override
-    public List<Cat> selectAllFromCatTable() {
-        List<Cat> cats = new ArrayList<>();
-        try (Connection connection = JdbcConnection.getConnection();
-             Statement statement = connection.createStatement();
-             ResultSet resultSet = statement.executeQuery(Queries.SELECT_FROM_CAT)) {
-
-            while (resultSet.next()) {
-                Cat cat = ResultSetToEntityConverter.convertToCat(resultSet);
-                cats.add(cat);
-            }
-        } catch (SQLException e) {
-            LOGGER.error(e.getMessage());
-        }
-        return cats;
-    }
-
-    @Override
-    public void dropTables() {
-        try (Connection connection = JdbcConnection.getConnection();
-             Statement statement = connection.createStatement()) {
-            statement.execute(Queries.DROP_TABLES);
-        } catch (SQLException exception) {
-            LOGGER.error("Error during dropping tables, {}", exception.getMessage());
-            throw new RepositoryException(exception.getMessage());
-        }
-    }
-
-    @Override
-    public void insertIntoTablesFromApi(List<CatDto> catsDto) {
-
-        try (Connection connection = JdbcConnection.getConnection()) {
-            catsDto.forEach(dto -> {
-                try {
-                    BreedDto breedDto = dto.getBreeds().getFirst();
-                    PreparedStatement selectFromBreedStatement = connection.prepareStatement(Queries.SELECT_BREED_ID_ID_FROM_BREED_WITH_BREED_ID);
-                    selectFromBreedStatement.setFetchSize(5);
-                    selectFromBreedStatement.setString(1, breedDto.getBreedId());
-                    ResultSet selectResulSet = selectFromBreedStatement.executeQuery();
-                    Integer breedId = null;
-                    if (!selectResulSet.next()) {
-                        PreparedStatement preparedStatement = connection.prepareStatement(Queries.INSERT_INTO_BREED, Statement.RETURN_GENERATED_KEYS);
-                        preparedStatement.setString(1, breedDto.getBreedId());
-                        preparedStatement.setString(2, breedDto.getName());
-                        preparedStatement.setString(3, breedDto.getTemperament());
-                        preparedStatement.setString(4, breedDto.getOrigin());
-                        preparedStatement.setString(5, breedDto.getCountryCode());
-                        preparedStatement.setString(6, breedDto.getDescription());
-                        preparedStatement.setString(7, breedDto.getLifeSpan());
-                        preparedStatement.setString(8, breedDto.getWikipediaUrl());
-                        preparedStatement.execute();
-                        ResultSet id = preparedStatement.getGeneratedKeys();
-                        if (id.next()) {
-                            breedId = id.getInt(1);
-                            breedDto.setId(breedId);
-                        }
-                        LOGGER.info("Breed with id = {}  was added to db", breedDto.getBreedId());
-                        preparedStatement.close();
-                    } else {
-                        breedId = selectResulSet.getInt("id");
-                    }
-                    PreparedStatement insertIntoCatStatement = connection.prepareStatement(Queries.INSERT_INTO_CAT);
-                    insertIntoCatStatement.setString(1, dto.getCatId());
-                    insertIntoCatStatement.setString(2, dto.getUrl());
-                    insertIntoCatStatement.setInt(3, breedId);
-                    insertIntoCatStatement.executeUpdate();
-                    LOGGER.info("Cat with id = {}  was added to db with breed = {}", dto.getCatId(), breedDto.getBreedId());
-                    selectFromBreedStatement.close();
-                    insertIntoCatStatement.close();
-                } catch (SQLException e) {
-                    LOGGER.error("Error during inserting Cat or Breed entities from API, {}", e.getMessage());
-                    throw new RepositoryException(e.getMessage());
-                }
-            });
-        } catch (SQLException e) {
-            LOGGER.error("Error during inserting values from API, {}", e.getMessage());
-            throw new RepositoryException(e.getMessage());
-        }
-    }
-
-    @Override
-    public DatabaseMetaData getMetaData() {
-        DatabaseMetaData metaData;
-        try (Connection connection = JdbcConnection.getConnection()) {
-            metaData = connection.getMetaData();
-        } catch (SQLException e) {
-            LOGGER.error("Error during getting metaData ,{}", e.getMessage());
-            throw new RepositoryException(e.getMessage());
-        }
-        return metaData;
-    }
-
-    @Override
-    public Optional<Breed> selectFromBreedById(Integer breedId) {
-        try (Connection connection = JdbcConnection.getConnection();
-             PreparedStatement prepStatement = connection.prepareStatement(Queries.SELECT_FROM_BREED_WITH_BREED_ID)) {
-            prepStatement.setInt(1, breedId);
-            ResultSet resultSet = prepStatement.executeQuery();
-            Breed breed = null;
-            if (resultSet.next()) {
-                breed = ResultSetToEntityConverter.convertToBreed(resultSet);
-            }
-            return Optional.ofNullable(breed);
-        } catch (SQLException exception) {
-            LOGGER.error("Error during selecting Breed by ID, {}", exception.getMessage());
-            throw new RepositoryException(exception.getMessage());
-        }
-    }
-
-    @Override
-    public Boolean updateCat(Cat cat) {
-        try (Connection connection = JdbcConnection.getConnection();
-             PreparedStatement prStatement = connection.prepareStatement(Queries.UPDATE_CAT)) {
-            prStatement.setString(1, cat.getCatId());
-            prStatement.setString(2, cat.getUrl());
-            prStatement.setInt(3, cat.getBreedId());
-            prStatement.setInt(4, cat.getId());
-            return prStatement.executeUpdate() > 0;
-        } catch (SQLException exception) {
-            LOGGER.error("Error during updating Cat with id={}, {}", cat.getId(), exception.getMessage());
-            throw new RepositoryException(exception.getMessage());
-        }
-    }
-
-    @Override
-    public Optional<List<Breed>> findAllBreedWithFilter(BreedFilter breedFilter) {
+    public List<Breed> findAllWithFilter(BreedFilter breedFilter) {
         List<Object> parameters = new ArrayList<>();
         List<String> whereSql = new ArrayList<>();
         if (breedFilter.getBreedId() != null) {
@@ -245,7 +58,7 @@ public class JdbcRepositoryImpl implements JdbcRepository {
         parameters.add(breedFilter.getOffset());
         String where = whereSql.stream().collect(Collectors.joining(
                 " AND ",
-                "WHERE ",
+                parameters.size() > 2 ? "WHERE " : "",
                 " LIMIT ? OFFSET ?"
         ));
         String selectAll = Queries.SELECT_ALL_FROM_BREED + where;
@@ -259,7 +72,7 @@ public class JdbcRepositoryImpl implements JdbcRepository {
             while (resultSet.next()) {
                 breeds.add(ResultSetToEntityConverter.convertToBreed(resultSet));
             }
-            return Optional.of(breeds);
+            return breeds;
         } catch (SQLException exception) {
             LOGGER.error("Error during selecting all Breed entities with filter, {}", exception.getMessage());
             throw new RepositoryException(exception.getMessage());
